[s8s] Move default reporter
